+++ Samples/JotFormTest/Classes/Controller/GetFormReportsViewController.m
@@ -73,7 +73,7 @@
         
         if ( result != nil ) {
             
-            int responseCode = [[result objectForKey:@"responseCode"] integerValue];
+            NSInteger responseCode = [[result objectForKey:@"responseCode"] integerValue];
             
             if ( responseCode == 200 || responseCode == 206 ) {
